fix some problem (#11)
* Update PlaceholderAPI.cpp

fix translate String API Error

* Update build.yml

* Update RegPAPI.cpp

Fixed floating-point number issue with player_hunger

* Update PlaceholderAPI.cpp

Fixed a bug with the translateString RemoteCall API

+++ BEPlaceholderAPI/RegPAPI.cpp
@@ -120,7 +120,7 @@ void regPlayerInit() {
 		});
 
 	PlaceholderAPI::registerPlayerPlaceholder("player_hunger", [](Player* sp) {
-		return S(sp->getAttribute(Player::HUNGER).getCurrentValue());
+		return S((int)sp->getAttribute(Player::HUNGER).getCurrentValue());
 		});
 	PlaceholderAPI::registerPlayerPlaceholder("player_max_hunger", [](Player* sp) {
 		string maxHunger = "";
